pyproject.toml
[tool.poetry]
name = "test-msio"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "msio", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = "^0.34.2"
structlog = "^25.3.0"
alembic = "^1.15.2"
asyncpg = "^0.30.0"
pydantic-settings = "2.9.1"
SQLAlchemy = "2.0.41"
passlib = "1.7.4"
python-jose = "3.5.0"







[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
